+++ server/src/services/graphql/mutation/user.ts
@@ -1,33 +1,48 @@
-import { GraphQLNonNull, GraphQLObjectType, GraphQLFieldConfigMap } from 'graphql';
+import { GraphQLFieldConfigMap, GraphQLNonNull, GraphQLObjectType } from 'graphql';
 import { GraphQLBoolean, GraphQLInt, GraphQLString } from 'graphql/type/scalars';
 import { BookType, DivisionType, UserType } from '../type/index';
 
 import { User, UserRepository } from '../../../repository/user/UserRepository';
 const userRepository = new UserRepository();
 
-const userMutation:GraphQLFieldConfigMap<any, any> = {
+const userMutation: GraphQLFieldConfigMap<any, any> = {
   addUser: {
     type: UserType,
     args: {
-      email: { type: new GraphQLNonNull(GraphQLString) },
+      id: { type: GraphQLInt },
+      email: { type: GraphQLString },
       name: { type: new GraphQLNonNull(GraphQLString) },
       birth: { type: new GraphQLNonNull(GraphQLString) },
       division: { type: new GraphQLNonNull(GraphQLInt) },
     },
-    resolve(parentValue, { email, birth, name, division }: User, context, info) {
+    resolve(parentValue, { id, email, birth, name, division }: User, context, info) {
+      if (!id && !email) {
+        return new Error('id or email is requirement.');
+      }
+      const dbUser = userRepository.findOne({ id, email });
+      if (dbUser) {
+        return new Error(`Already '${email}' is existed.`);
+      }
       return userRepository.create({ email, birth, name, division });
     },
   },
   editUser: {
     type: UserType,
     args: {
-      id: { type: new GraphQLNonNull(GraphQLInt) },
-      email: { type: new GraphQLNonNull(GraphQLString) },
+      id: { type: GraphQLInt },
+      email: { type: GraphQLString },
       name: { type: new GraphQLNonNull(GraphQLString) },
       birth: { type: new GraphQLNonNull(GraphQLString) },
       division: { type: new GraphQLNonNull(GraphQLInt) },
     },
-    resolve(parentValue, { birth, name, division }: User) {
+    resolve(parentValue, { id, email, birth, name, division }: User) {
+      if (!id && !email) {
+        return new Error('id or email is requirement.');
+      }
+      const dbUser = userRepository.findOne({ id, email });
+      if (!dbUser) {
+        return new Error('The user is not found');
+      }
       return userRepository.update({ birth, name, division });
     },
   },
@@ -38,9 +53,16 @@ const userMutation:GraphQLFieldConfigMap<any, any> = {
       email: { type: GraphQLString },
     },
     resolve(parentValue, { id, email }: User) {
-      return userRepository.delete({id, email});
+      if (!id && !email) {
+        return new Error(`id or email is requirement.`);
+      }
+      const dbUser = userRepository.findOne({ id, email });
+      if (!dbUser) {
+        return new Error('The user is not found');
+      }
+      return userRepository.delete({ id, email });
     },
   },
-}
+};
 
 export default userMutation;
